Refactor: nested resources under api/v1/
