+++ config/routes.rb
@@ -4,9 +4,11 @@ Rails.application.routes.draw do
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
   # Can be used by load balancers and uptime monitors to verify that the app is live.
   get "up" => "rails/health#show", as: :rails_health_check
-
-  resources :subjects
-  resources :notes
-
-  resources :subject_notes, only: [:create, :destroy]
+  namespace :api do
+    namespace :v1 do
+      resources :subjects, only: [:create, :index, :show, :update, :destroy]
+      resources :notes, only: [:create, :index, :show, :update, :destroy]
+      resources :subject_notes, only: [:create, :index, :show, :update, :destroy]
+    end
+  end
 end
